package/docker-compose: bump to version 2.10.0
Docker Compose v2 is no longer a standalone component, but is now a
plugin loaded by docker-cli.

As such, it should not be installed in /usr/bin, but in the directory
where docker-cli loads its plugins from.

Additionally, we consequently make docker-compose depend on docker-cli;
indeed, it does not really make sense to present a plugin unless the
component it attaches to is already enabled [0].

License hash changed due to strictly copying the license text template,
without customisation to the year and copyright owner.

[0] the original submission by Christian would use a select, to keep
existing config, but that's not sensible, as we already have some
packages that are plugins and that use depends-on, like nginx plugins.
For consistency and as it semantically makes sense, we use a depends-on
here too.

Signed-off-by: Christian Stewart <christian@paral.in>
[yann.morin.1998@free.fr:
  - don't select docker-cli, but depends-on it; explain it in commit log
  - explain why we override the install commands
  - explain change in license file hash
]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>